requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
psycopg2-binary
sqlalchemy
asyncpg
python-dotenv